make the no-code actually generic
