+++ ui/src/components/flows/noCodeTypes.ts
@@ -1,0 +1,9 @@
+export interface NoCodeProps {
+    creatingTask?: boolean;
+    editingTask?: boolean;
+    parentPath?: string;
+    refPath?: number;
+    position?: "before" | "after";
+    blockSchemaPath?: string;
+    fieldName?: string | undefined;
+}
